+++ packages/components/modules/comments/Comments/__tests__/__utils__/CommentsForTesting/index.tsx
@@ -2,9 +2,9 @@ import { graphql, useLazyLoadQuery } from 'react-relay'
 
 import { CommentsForTestingQuery } from '../../../../../../__generated__/CommentsForTestingQuery.graphql'
 import { CommentsFragment$key } from '../../../../../../__generated__/CommentsFragment.graphql'
+import withProviders from '../../../../../../__test_utils__/withProviders'
 import Comments from '../../../index'
 import { CommentsProps } from '../../../types'
-import withProviders from '../withProviders'
 
 const CommentsForTesting = (props?: Partial<CommentsProps>) => {
   const data = useLazyLoadQuery<CommentsForTestingQuery>(