+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseapp-frontend/utils
 
+## 3.0.2
+
+### Patch Changes
+
+- `getCookie` and `getCookieAsync` won't parse the value as default.
+- `setCookie` won't stringfy the value as default.
+
 ## 3.0.1
 
 ### Patch Changes